+++ usecase/src/main/kotlin/dev/agner/portfolio/usecase/bond/model/BondOrder.kt
@@ -1,6 +1,7 @@
 package dev.agner.portfolio.usecase.bond.model
 
 import kotlinx.datetime.LocalDate
+import java.math.BigDecimal
 
 enum class BondOrderType {
     BUY,
@@ -12,12 +13,12 @@ data class BondOrder(
     val bond: Bond,
     val type: BondOrderType,
     val date: LocalDate,
-    val amount: Double,
+    val amount: BigDecimal,
 )
 
 data class BondOrderCreation(
     val bondId: Int,
     val type: BondOrderType,
     val date: LocalDate,
-    val amount: Double,
+    val amount: BigDecimal,
 )
